requirements.txt
streamlit>=1.30.0
Pillow>=10.2.0
numpy>=1.26.0
streamlit-drawable-canvas>=0.8.0